Add first container test
This test verifies a very simple use case of the container. Namely
that it can scan a fixed lib path, extract from it services that
implement interfaces, and register those services.
It will also prove that it can instance a singleton service and
that it fetches the correct implementation class for it.