Add massive refactor to dependency building
To enable YAML config file to configure injection points on
attributes of injected services, a massive refactor was needed
on how to build the dependencies and how to configure them, both
on the code config and the file config manner.

Now there's more flexibility on how you can configure your services.

You won't need anymore the injected trait if the default values are
enough for your service's dependencies. It will only be used if you
want request-injected dependencies.

Now the Moose service factory will automatically try to build every
dependency of a service for which an entry is found in the registry.
If the dependency is required and the service is not found, an error
will be thrown. Otherwise, we can assume it is not important.